requirements.txt
# Quran AI Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai-whisper>=20231117
torch>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
transformers>=4.35.0
rapidfuzz>=3.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0

# Optional: Diacritization (Seerat AI mode). Uncomment one:
# camel-tools  # CAMeL Tools
# transformers already above; for AraT5 use: UBC-NLP/AraT5-base-diacritization
